Fix w64 build
* src/ribbon-callbacks.cpp (launch_cert_details): Cast HWND w64
compatible.

+++ src/ribbon-callbacks.cpp
@@ -1740,7 +1740,7 @@ HRESULT launch_cert_details (LPDISPATCH ctrl)
               log_error ("%s:%s: No spawn engine.",
                          SRCNAME, __func__);
             }
-            std::string parentWid = std::to_string ((unsigned int) get_active_hwnd ());
+            std::string parentWid = std::to_string ((int) (intptr_t) get_active_hwnd ());
             const char *argv[] = {path.c_str(),
                                   "--query",
                                   mail->get_sig_fpr(),
